Move socket, slot and networking to shared [WIP]
