+++ packages/shared/pkg/instance/slot.go
